Modify : PongOneConsumers in game/PongConsumers.py
