+++ srcs/requirements/django/src/game/consumers/PongConsumers.py
@@ -1,9 +1,8 @@
-import json, logging
-
+import json, logging, asyncio
 from channels.generic.websocket import AsyncWebsocketConsumer
 from channels.layers import get_channel_layer
-from users.models import User, UserRecordFightingGame
-from game.info_fight import Room
+from users.models import User, UserRecordPongGame
+from game.info_pong import Bar, Player, Team, Room
 from game.models import GameRoom
 from game.utils import rating_calculator
 from datetime import datetime
@@ -14,31 +13,30 @@ from urllib.parse import parse_qs
 
 logger = logging.getLogger(__name__)
 
-class fightingConsumers(AsyncWebsocketConsumer):
+class PongOneConsumers(AsyncWebsocketConsumer):
     class RoomList:
         pass
 
+    # 웹소켓에서 연결 요청이 들어오면 받고 상대방 매칭 시도
     async def connect(self):
         self.game_group_name = ""
-        self.rating_difference = 100
+        self.rating_differece = 100
         self.create_time = datetime.now()
         query_string = parse_qs(self.scope['query_string'].decode())
-        access_token = query_string.get('access', None)[0]
+        access_token = query_string.get('access_token', None)[0]
         self.user_name = access_token_get_name(access_token)
 
         if access_token == None or self.user_name == None:
             await self.close()
         
-        self.display_name = await self.get_display_name()
-        self.rating = await self.get_rating(self.user_name)
+        self.rating =  await self.get_rating(self.user_name)
         await self.channel_layer.group_add("game_queue", self.channel_name)
         await self.accept()
         await self.send(text_data=json.dumps({
             "type" : "game.message",
             "data" : {
                 "mode" : "connect",
-                "name" : self.user_name,
-                "display_name" : self.display_name,
+                "name" : self.channel_name,
             }
         }))
         await self.join_matching()
@@ -50,7 +48,8 @@ class fightingConsumers(AsyncWebsocketConsumer):
                 await delattr(self.RoomList, self.game_group_name)
                 await self.db_delete()
         except AttributeError:
-            logger.debug('No room')
+            logger.debug("No room")
+
         if close_code == 1000:
             if self.game_group_name:
                 await self.channel_layer.group_send(
@@ -62,10 +61,13 @@ class fightingConsumers(AsyncWebsocketConsumer):
                     }
                 )
                 await self.channel_layer.group_discard(self.game_group_name, self.channel_name)
+            room = self.get_room()
+            if room != None:
+                delattr(self.RoomList, self.game_group_name)
         else:
-            logger.error("websocket " + self.channel_name + ": abnormal.termination")
+            logger.error("websocket " + self.channel_name + ": abnormal termination")
             if self.game_group_name:
-                await self.channel_layer.group_send(
+                self.channel_layer.group_send(
                     self.game_group_name, {
                         'type' : 'game.message',
                         'data' : {
@@ -79,13 +81,14 @@ class fightingConsumers(AsyncWebsocketConsumer):
         data = json.loads(text_data)
         msg_type = data.get('type')
         msg_data = data.get('data', [])
+        # 연결 후에 웹소켓에서 플레이어 정보를 보낸다
         if msg_type == "set.game":
             cnt = await self.db_cnt()
             room = await self.get_room()
             if cnt == 1 and room == None:
                 setattr(self.RoomList, self.game_group_name, Room())
                 room = await self.get_room()
-                room.setPlayer({"name": msg_data['player0'], "rating": 0}, {"name": msg_data['player1'], "rating": 0})
+                room.setPlayer({"name": msg_data['player0'], "rating": msg_data['player0Rating']}, {"name": msg_data['player1'], "rating": msg_data['player1Rating']})
             if cnt == 2 and room != None:
                 await self.channel_layer.group_send(
                     self.game_group_name, {
@@ -95,67 +98,25 @@ class fightingConsumers(AsyncWebsocketConsumer):
                         }
                     }
                 )
-        if msg_type == 'select.info':
+        # 각 플레이어들의 탁구채 위치 정보. 정보를 받으면 최신화
+        if msg_type == 'bar.info':
             room = await self.get_room()
-            if room.player0.name == msg_data['name']:
-                room.player0.fighter = msg_data['select']['player0']
-                fighter = msg_data['select']['player0']
+            if room.player0.name == self.channel_name:
+                room.player0.bar.x = msg_data['bar']['x']
+                room.player0.bar.y = msg_data['bar']['y']
             else:
-                room.player1.fighter = msg_data['select']['player1']
-                fighter = msg_data['select']['player1']
-            await self.update_select_info(fighter)
-        if msg_type == "set.select":
-            cnt = await self.db_cnt()
-            room = await self.get_room()
-            if cnt == 2:
-                await self.channel_layer.group_send(
-                    self.game_group_name, {
-                        "type" : "game.message",
-                        "data" : {
-                            "mode" : "complete.select",
-						}
-					}
-				)
-        if msg_type == "battle.info":
-            room = await self.get_room()
-            if room.player0.name == msg_data['name']:
-                room.player0.character.updateInfo(msg_data['x'], msg_data['y'], msg_data['state'], msg_data['health'])
-            else:
-                room.player1.character.updateInfo(msg_data['x'], msg_data['y'], msg_data['state'], msg_data['health'])
-            await self.update_battle_info({"x" : msg_data['x'], "y" : msg_data['y'], "state" : msg_data['state'], "health" : msg_data['health']})  
-        if msg_type == "battle.complete":
-            cnt = await self.db_cnt()
-            room = await self.get_room()
-            if room.winner == "":
-                room.winner = msg_data['winner']
-                await self.calculate_rating()
-            if cnt == 2:
-                await self.channel_layer.group_send(
-                    self.game_group_name, {
-                        "type" : "game.message",
-                        "data" : {
-                            "mode" : "complete.battle",
-                        }
-                    }
-                )
-        if msg_type == "result.complete":
-            cnt = await self.db_cnt()
-            room = await self.get_room()
-            if cnt == 2:
-                await self.channel_layer.group_send(
-                    self.game_group_name, {
-                        "type" : "game.message",
-                        "data" : {
-                            "mode" : "complete.result",
-                        }
-                    }
-                )
+                room.player1.bar.x = msg_data['bar']['x']
+                room.player1.bar.y = msg_data['bar']['y']
+            await self.update_bar_info()
+        # 게임이 전부 끝나면 종료한다
         if msg_type == "game.clear":
-            self.disconnect(1001)
+            self.disconnect(1000)
 
+    # 매칭 시도
     async def join_matching(self):
         count = await self.get_room_cnt()
         name = await self.get_room_name()
+        # 방이 하나도 안 만들어졌거나 전부 게임중이면 방을 새로 판다. 방이 있으면 해당 방에 참여
         if count == 0 or name == "not":
             await self.create_room()
             await self.channel_layer.group_add(self.game_group_name, self.channel_name)
@@ -169,50 +130,43 @@ class fightingConsumers(AsyncWebsocketConsumer):
                     await self.set_player(1)
                     player = await self.get_player()
                     await self.channel_layer.group_send(
-						self.game_group_name, {
-							"type" : 'game.message',
-							"data" : {
-								"mode" : "set.game",
-								"player0" : player["player0"],
-                                "player0display" : player["player0display"],
-								"player1" : player["player1"],
-                                "player1display" : player["player1display"],
-								"group" : self.game_group_name,
-							}
-						}
-					)
-            else:
-                if self.matching_timeout():
-                    await self.close()
-                self.rating_difference += 200
-                await self.join_matching()
-    
+                        self.game_group_name, {
+                            "type" : 'game.message',
+                            "data" : {
+                                "mode" : "set.game",
+                                "player0" : player["player0"],
+                                "player1" : player["player1"],
+                                "group" : self.game_group_name,
+                            }
+                        }
+                    )
+                else:
+                    if self.matching_timeout():
+                        await self.close()
+                    self.rating_differece += 200
+                    await self.join_matching()
+
+    # 본인이 속한 방을 불러온다. 없으면 None
     async def get_room(self):
         return getattr(self.RoomList, self.game_group_name, None)
-
-    async def update_select_info(self, fighter):
+    
+    # 사용자에게 받은 탁구채 위치 정보를 최신화
+    async def update_bar_info(self):
+        room = await self.get_room()
+        if room.player0.name == self.channel_name:
+            bar = {"x" : room.player0.bar.x, "y" : room.player0.bar.y}
+        elif room.player1.name == self.channel_name:
+            bar = {"x" : room.player1.bar.x, "y" : room.player1.bar.y}
         await self.channel_layer.group_send(
             self.game_group_name, {
                 "type" : 'game.message',
                 "data" : {
-                    "mode" : "update.select",
-                    "name" : self.user_name,
-                    "fighter" : fighter,
-                }
-            }
-        )
-    
-    async def update_battle_info(self, info):
-        await self.channel_layer.group_send(
-            self.game_group_name, {
-                "type" : "game.message",
-                "data" : {
-                    "mode" : "update.battle",
-                    "name" : self.user_name,
-                    "x" : info["x"],
-                    "y" : info["y"],
-                    "state" : info["state"],
-                    "health" : info["health"],
+                    "mode" : "update.bar",
+                    "name" : self.channel_name,
+                    "bar" : {
+                        "x" : bar['x'],
+                        "y" : bar['y']
+                    }
                 }
             }
         )
@@ -222,7 +176,7 @@ class fightingConsumers(AsyncWebsocketConsumer):
         if difference.seconds > 10:
             return False
         return True
-
+    
     async def get_group_member_count(group_name):
         channel_layer = get_channel_layer()
         group_info = await channel_layer.group_layer.group_status(group_name)
@@ -231,9 +185,10 @@ class fightingConsumers(AsyncWebsocketConsumer):
         else:
             return -1
 
+    # 웹소켓에 보낼 메세지들을 처리한다
     async def game_message(self, event):
         await self.send(text_data=json.dumps(event))
-    
+
     async def calculate_rating(self):
         is_room = getattr(self.RoomList, self.game_group_name, None)
         if is_room.winner == is_room.player0.name:
@@ -244,6 +199,7 @@ class fightingConsumers(AsyncWebsocketConsumer):
             player1rating = rating_calculator(is_room.player1.rating, is_room.player0.rating, 1)
         await self.set_rating([is_room.player0.name, player0rating], [is_room.player1.name, player1rating], is_room.winner)
 
+    # database를 이용하여 사람 수를 세거나 할 때 사용
     @database_sync_to_async
     def db_cnt(self):
         is_room = GameRoom.objects.get(room_name=self.game_group_name)
@@ -256,21 +212,19 @@ class fightingConsumers(AsyncWebsocketConsumer):
         is_room.save()
         return cnt
 
+    # 게임방이 얼마나 있는 지 확인할 때 사용
     @database_sync_to_async
     def get_room_cnt(self):
         return GameRoom.objects.count()
     
-    @database_sync_to_async
-    def get_display_name(self):
-        is_user = User.objects.get(username=self.user_name)
-        return is_user.display_name
-    
+    # 방을 만든다. 방 이름은 "pong_one_" + 랜덤 생성 아스키 코드 6자리
     @database_sync_to_async
     def create_room(self):
         self.game_group_name = self.user_name + random_key(6)
-        is_room = GameRoom(room_name=self.game_group_name, mode="fighting", status="waiting", player0=self.user_name, player0displayName=self.display_name)
+        is_room = GameRoom(room_name=self.game_group_name, mode="pingpong", status="waiting", player0=self.channel_name)
         is_room.save()
     
+    # 들어갈 수 있는 방을 검색하고 들어간 방의 이름을 알려준다
     @database_sync_to_async
     def get_room_name(self):
         room_all = GameRoom.objects.all()
@@ -282,47 +236,47 @@ class fightingConsumers(AsyncWebsocketConsumer):
                 return room.name
         return "not"
     
+    # 게임방에 두번째 사용자 등록
     @database_sync_to_async
     def set_player(self, player_num):
         is_room = GameRoom.objects.get(room_name=self.game_group_name)
         if player_num == 0:
             is_room.player0 = self.user_name
-            is_room.player0displayName = self.display_name
         else:
             is_room.player1 = self.user_name
-            is_room.player1displayName = self.display_name
         is_room.save()
     
+    # 게임방에 속한 플레이어들을 불러온다
     @database_sync_to_async
     def get_player(self):
         is_room = GameRoom.objects.get(room_name=self.game_group_name)
-        return {"player0": is_room.player0, "player0display" : is_room.player0displayName, "player1": is_room.player1, "player1display" : is_room.player1displayName}
-
+        return {"player0": is_room.player0, "player1": is_room.player1}
+    
     @database_sync_to_async
     def db_delete(self):
         is_room = GameRoom.objects.get(room_name=self.game_group_name)
         is_room.delete()
-    
+
     @database_sync_to_async
     def get_rating(self, name):
         logger.error(name)
         is_user = User.objects.get(username=name)
-        record = UserRecordFightingGame.objects.get(me=is_user)
+        record = UserRecordPongGame.objects.get(me=is_user)
         return record.rating
-
+    
     @database_sync_to_async
     def rating_check(self, group_name):
         is_room = GameRoom.objects.get(room_name=group_name)
-        if is_room.player0Rating > self.rating + self.rating_difference or is_room.player0Rating < self.rating - self.rating_difference:
+        if is_room.player0rating > self.rating + self.rating_differece or is_room.player0rating < self.rating - self.rating_differece:
             return False
         return True
     
     @database_sync_to_async
     def set_rating(self, player0Info, player1Info, winner):
         player0 = User.objects.get(username=player0Info[0])
-        player0record = UserRecordFightingGame.objects.get(me=player0)
+        player0record = UserRecordPongGame.objects.get(me=player0)
         player1 = User.objects.get(username=player1Info[0])
-        player1record = UserRecordFightingGame.objects.get(me=player1)
+        player1record = UserRecordPongGame.objects.get(me=player1)
         player0record.rating = player0Info[1]
         player1record.rating = player1Info[1]
         if winner == player0.username:
@@ -334,3 +288,6 @@ class fightingConsumers(AsyncWebsocketConsumer):
         player0record.save()
         player1record.save()
 
+# class PongTwoConsumers(AsyncWebsocketConsumer):
+
+# class PongTournamentConsumers(AsyncWebsocketConsumer):
