+++ srcs/requirements/front/srcs/games/entities/fighters/playerOne.js
@@ -9,3 +9,6 @@ export class PlayerOne extends Fighter {
 		this.bar.COLOR = "#DB4455";
 	}
 }
+
+// this.bar.Y = view.HEIGHT / 4 - this.bar.HEIGHT / 2;
+// this.bar.Y = view.HEIGHT / 4 * 3 - this.bar.HEIGHT / 2;
